Primera versión completa y funcionando.

+++ src/Moneda.java
@@ -1,4 +1,4 @@
 public record Moneda(String base_code,
                      String target_code,
-                     int conversion_result) {
+                     double conversion_result) {
 }
